+++ src/components/MegaGallery/components/FolderTree/FolderTreePreview.vue
@@ -175,7 +175,7 @@ export default {
 					notAllowedIds: [],
 				},
 			],
-		}
+		};
 	},
 	computed: {
 		...mapState({