+++ src/components/MegaGallery/components/Settings.vue
@@ -41,6 +41,84 @@
 				</v-col>
 			</v-row>
 		</OptionsWrapper>
+
+		<h3>Panels</h3>
+		<p>
+			You can control what panels are visible for the users.
+		</p>
+		<OptionsWrapper>
+			<v-card elevation="0" class="d-flex px-2">
+				<v-card
+					min-height="72px"
+					flat
+					class="rounded-0 rounded-t d-flex pa-4"
+					width="100%"
+				>
+					<v-row>
+						<v-col
+							v-for="(value, name) in panels"
+							:key="name"
+							cols="6"
+							class="d-flex justify-space-between align-center"
+						>
+							<v-card-title
+								class="ma-0 pa-0 text-subtitle-1"
+								style="margin-bottom: -3px !important"
+							>
+								{{ name }}
+							</v-card-title>
+							<v-switch
+								:input-value="panels[name]"
+								class="ma-0 pa-0"
+								color="primary"
+								hide-details
+								:true-value="true"
+								@change="updatePanel($event, name)"
+							/>
+						</v-col>
+					</v-row>
+				</v-card>
+			</v-card>
+		</OptionsWrapper>
+
+		<h3>Upload options</h3>
+		<p>
+			You can control what upload options are visible for the users.
+		</p>
+		<OptionsWrapper>
+			<v-card elevation="0" class="d-flex px-2">
+				<v-card
+					min-height="72px"
+					flat
+					class="rounded-0 rounded-t d-flex pa-4"
+					width="100%"
+				>
+					<v-row>
+						<v-col
+							v-for="(value, name) in uploadOptions"
+							:key="name"
+							cols="6"
+							class="d-flex justify-space-between align-center"
+						>
+							<v-card-title
+								class="ma-0 pa-0 text-subtitle-1"
+								style="margin-bottom: -3px !important"
+							>
+								{{ name }}
+							</v-card-title>
+							<v-switch
+								:input-value="uploadOptions[name]"
+								class="ma-0 pa-0"
+								color="primary"
+								hide-details
+								:true-value="true"
+								@change="updateUploadOption($event, name)"
+							/>
+						</v-col>
+					</v-row>
+				</v-card>
+			</v-card>
+		</OptionsWrapper>
 	</div>
 </template>
 
@@ -75,16 +153,32 @@ export default {
 				this.updateGallerySettings();
 			},
 		},
+		panels() {
+			return this.$store.state.megaGalleryConfig.settings.panels;
+		},
+		uploadOptions() {
+			return this.$store.state.megaGalleryConfig.settings.uploadOptions;
+		},
 
 	},
 	methods: {
 		...mapMutations({
 			setMaxFileSize: "setMaxFileSize",
 			setMaxImagePerPage: "setMaxImagePerPage",
+			updateGalleryUploadOption: "updateGalleryUploadOption",
+			updateGalleryPanel: "updateGalleryPanel",
 		}),
 		...mapActions({
 			updateGallerySettings: "updateGallerySettings",
 		}),
+		updatePanel(value, name) {
+			this.updateGalleryPanel({ [name]: !!value });
+			this.updateGallerySettings();
+		},
+		updateUploadOption(value, name) {
+			this.updateGalleryUploadOption({ [name]: !!value });
+			this.updateGallerySettings();
+		},
 	},
 };
 </script>